content: Introduce message-bubble widget
This allowed to simplify message-row by a lot.
