+++ data/resources/resources.gresource.xml
@@ -5,6 +5,7 @@
     <file compressed="true" preprocess="xml-stripblanks">ui/content.ui</file>
     <file compressed="true" preprocess="xml-stripblanks">ui/content-chat-history.ui</file>
     <file compressed="true" preprocess="xml-stripblanks">ui/content-event-row.ui</file>
+    <file compressed="true" preprocess="xml-stripblanks">ui/content-message-bubble.ui</file>
     <file compressed="true" preprocess="xml-stripblanks">ui/content-message-row.ui</file>
     <file compressed="true" preprocess="xml-stripblanks">ui/login.ui</file>
     <file compressed="true" preprocess="xml-stripblanks">ui/session.ui</file>
